+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to Xolace are documented here.
 
 ---
 
-## [Unreleased]
+## [1.7.0] - (2026-07-14)
 
 ### Added